refactor(payment): PAYPAL-1538 switched PayPalCommerce common button strategy with V2 and removed old implementation
