+++ packages/core/src/checkout-buttons/checkout-button-options.ts
@@ -6,7 +6,7 @@ import { ApplePayButtonInitializeOptions } from './strategies/apple-pay';
 import { BraintreePaypalButtonInitializeOptions, BraintreePaypalCreditButtonInitializeOptions, BraintreeVenmoButtonInitializeOptions } from './strategies/braintree';
 import { GooglePayButtonInitializeOptions } from './strategies/googlepay';
 import { PaypalButtonInitializeOptions } from './strategies/paypal';
-import { PaypalCommerceAlternativeMethodsButtonOptions, PaypalCommerceButtonInitializeOptions, PaypalCommerceCreditButtonInitializeOptions, PaypalCommerceInlineCheckoutButtonInitializeOptions, PaypalCommerceV2ButtonInitializeOptions, PaypalCommerceVenmoButtonInitializeOptions } from './strategies/paypal-commerce';
+import { PaypalCommerceAlternativeMethodsButtonOptions, PaypalCommerceButtonInitializeOptions, PaypalCommerceCreditButtonInitializeOptions, PaypalCommerceInlineCheckoutButtonInitializeOptions, PaypalCommerceVenmoButtonInitializeOptions } from './strategies/paypal-commerce';
 
 export { CheckoutButtonInitializeOptions } from '../generated/checkout-button-initialize-options';
 
@@ -122,13 +122,14 @@ export interface BaseCheckoutButtonInitializeOptions extends CheckoutButtonOptio
      * The options that are required to facilitate PayPal Commerce. They can be omitted
      * unless you need to support Paypal.
      */
+    // TODO: should be removed when PAYPAL-1539 hits Tier3
     paypalCommerce?: PaypalCommerceButtonInitializeOptions;
 
     /**
      * The options that are required to facilitate PayPal Commerce V2. They can be omitted
      * unless you need to support Paypal Commerce.
      */
-    paypalcommerce?: PaypalCommerceV2ButtonInitializeOptions;
+    paypalcommerce?: PaypalCommerceButtonInitializeOptions;
 
     /**
      * The options that are required to facilitate PayPal Commerce. They can be omitted
