+++ packages/core/src/checkout-buttons/checkout-button-strategy-action-creator.ts
@@ -20,10 +20,12 @@ import {
 import CheckoutButtonRegistryV2 from "./checkout-button-strategy-registry-v2";
 import { CheckoutButtonMethodType, CheckoutButtonStrategy } from "./strategies";
 
+// TODO: should be removed when PAYPAL-1539 hits Tier3
 const methodMap: { [key: string]: CheckoutButtonMethodType | undefined } = {
-    [CheckoutButtonMethodType.PAYPALCOMMERCEV2]: CheckoutButtonMethodType.PAYPALCOMMERCE, // TODO: should be removed after PaypalCommerceButtonStrategy deprecation
+    [CheckoutButtonMethodType.PAYPALCOMMERCEV2]: CheckoutButtonMethodType.PAYPALCOMMERCE,
 };
 
+// TODO: should be removed when PAYPAL-1539 hits Tier3
 const mapCheckoutButtonMethodId = (methodId: CheckoutButtonMethodType) => {
     return methodMap[methodId] || methodId;
 };
@@ -64,7 +66,7 @@ export default class CheckoutButtonStrategyActionCreator {
                     )
                 ),
                 this._paymentMethodActionCreator.loadPaymentMethod(
-                    mapCheckoutButtonMethodId(options.methodId),
+                    mapCheckoutButtonMethodId(options.methodId), // TODO: the line should be updated with 'options.methodId,' when PAYPAL-1539 hits Tier3
                     { timeout: options.timeout, useCache: true }
                 )(store),
                 defer(() =>
