--- packages/core/src/checkout-buttons/strategies/paypal-commerce/paypal-commerce-v2-button-options.ts
@@ -1,13 +1,0 @@
-import { PaypalButtonStyleOptions } from '../../../payment/strategies/paypal-commerce';
-
-export interface PaypalCommerceV2ButtonInitializeOptions {
-    /**
-     * A set of styling options for the checkout button.
-     */
-    style?: PaypalButtonStyleOptions;
-
-    /**
-     * Flag which helps to detect that the strategy initializes on Checkout page
-     */
-    initializesOnCheckoutPage?: boolean;
-}
